Prepare support to Asciigrid (ESRI and GRASS). Thanks to image-io ext
