+++ pom.xml
@@ -42,6 +42,7 @@
         <sl4j-version>1.6.0</sl4j-version>
         <cts-version>1.3.3</cts-version>
         <junit-version>4.12</junit-version>
+        <imageio-ext-arcgrid-version>1.2-SNAPSHOT</imageio-ext-arcgrid-version>
     </properties>
     <scm>
         <connection>scm:git:https://github.com/orbisgis/H2GIS.git</connection>
@@ -115,6 +116,14 @@
             <id>repo2.maven.org</id>
             <name>Maven2 repository 2</name>
             <url>http://repo2.maven.org/maven2</url>
+        </repository>
+        <repository>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
+            <id>imageio-ext-repository</id>
+            <name>imageio-ext Repository</name>
+            <url>http://maven.geo-solutions.it/</url>
         </repository>
     </repositories>
     <parent>
